python/src/retired/_requirements.txt
-r torch_cpu.txt
numpy>=1.24,<2.4
numba>=0.63,<0.64
transformers
datasets
tqdm
beautifulsoup4
xmltodict
fasttext-wheel
matplotlib
seaborn
psycopg
pyright
mypy
ruff
lxml
psycopg-binary
types-tqdm
scikit-learn-stubs
Levenshtein
faiss-cpu
flask
flask-cors
waitress
nltk
plotly
umap-learn
graphviz
hdbscan
dash
